Fix errors in app.py, and refactor tests folder
